Fixed a bug in MimeParser.StepHeaders()
If we don't advance any bytes this loop, don't set midline to true.

Fixes a case where the '\r' of a blank line that terminates the headers
is the last byte of the previous read(). If we set midline to true, then
what happens is that we assume that the blank line was part of the headers
and never stop scanning the headers until we reach the end of the file.

Thanks to Hugo Hallqvist for finding & reporting this bug.