Fixed MessageIdList to not be too strict in what it accepts
Since we have no control over what the formatting is for messages
that we are parsing, be very liberal in what we accept.

This was causing at least one user problems when parsing a
MimeMessage that had a References header with a msgid that looked
like <xxxxx.yyyyy@>. EnumerateReferences() handled this correctly,
but MessageIdList.Add() rejected it.