Added two overloads that don't require a name;.
In many cases I don't care about an actual name; when some monitoring or logging system or a 'bot' is sending email there is no 'real name' anyway - setting the name equal to the address is fine (for me) in that situation. This commit adds two 'convenience constructors' which do away with the name argument and set the name to the given address by calling the overloaded constructors with the address value for the name argument.