Modified Rfc2047's internal Token class to be a struct
This reduces the number of allocations needed, but it does mean that
List<Token> will be much larger.

To combat this a bit, redesigned the Token class/struct to have fewer
fields which reduced sizeof(Token) from 40 bytes down to 24.

Ideally, we'd get that down even further to 16 bytes (as recommended by
Microsoft). Added some comments with ideas on how to accomplish that.