Improve Header.GetRawValue() logic
Now that users can modify FormatOptions.Default, we need to be smarter
about whether or not we need to reformat header values in
Header.GetRawValue().

The old logic used to assume that cached raw values would always use
FormatOptions.International=False unless they were explicitly set, but
now that FormatOptions.Default.International can be True, we need to cache
that value and determine if it has changed since the rawValue was last
cached.

Realistically, we should be caching other FormatOptions values as well,
but this was the most important one (and the only one that was used to
determine whether the rawValue needed to be reformatted or not).

Also modified the ARC and DKIM verifier logic to share code that clones
the user-requested FormatOptions to override certain values like
NewLineFormat to also set VerifyingSignature = True.

This *may* solve issue #1000