Revert change making it possible to change FormatOptions.Default property values
This reverts the following commits:
* 4b77a9f6607d722f3f0527ff0315261d53d87329: Allow changing the FormatOptions.Default values
* e3bab7b5efb153a32b67a873227f5ec13230c1f7: Improve Header.GetRawValue() logic
* 22881de2d361e3464d4a9038fdb5cf3c9cca2c9e: Cache more of the FormatOptions in each Header
* 9570ebd54ac25af202242795ad3132ed705b3954: Added FormatOptions.ReformatHeaders and fixed unit tests

...And reopens issue #993.

I will be revisiting this set of changes for v5.0 when I'll have more flexibility with
behavioral changes. The problem with this set was not that it "didn't work", but rather
that it changed expected behavior. I also wasn't sure if I really liked the new
FormatOptions.ReformatHeaders property naming.