Modified AttachmentCollection.Add()
When adding a message/rfc822 stream, if the content fails to parse
as a MimeMessage, then fall back to attaching the content as
application/octet-stream.

This is only possible if the stream is seekable and if the mime-type
was auto-detected. If the user has specifically requested that we
attach as a message/rfc822, allow the FormatException to bubble back up
to our caller.

Fixes issue #1001