Added support for addresses like "webmaster\@custom-host.com@mail-host.com"
If we encounter a \@ sequence, convert that to %40 when the
FormatOptions.AddressParserComplianceMode value is Looser.

This solution isn't ideal, but is probably the simplest option that we
can do for invalid local-parts like this.

The other option would be to quote the local-part, but that would be a
much more involved fix because obs-local-part allows a mix of qstring
and atom tokens separated by '.' in a local-part (the modern form only
allows a single qstring -or- multiple atoms separated by '.'s).

Because of obs-local-part, we can't just wrap the token with DQUOTEs
when we finish consuming the local-part, because it *could* include
1 or more qstrings that we would have to escape. We also can't just
quote individual atom tokens containing the \@ sequence, because then
the esxample address above would end up being:

    "webmaster@custom-host".com

Even though that would be syntactically valid, it's not likely to be
interpreted the same. Ideally, if we were to implement a solution that
quoted the relevant parts of the local-part token, it would look like
this:

    "webmaster@custom-host.com"

This is *doable*, but not without significant rewriting of the current
TryParseLocalPart method logic.

That said, even *that* might not get interpreted the as the same mailbox
by whatever mail software generated the \@ sequence in the first place.

(Obviously, the same goes for this %40 hack.)

Fixes issue #1043