Improve InternetAddress[List] parser performance for malformed addresses
In the case described in issue #1106, there are 1000's of "addresses"
which are really only quoted-strings separated by commas.

What this patch does is to prevent the parser from consuming the entire
string as 1 display-name, then deciding that there's no address (reached
end of string), so falling back to parsing the first qstring as a local-part
instead.

This was causing the parser to be O(N^2), so the larger that recipient list
got, the worse the performance got.

Fixes issue #1106