Make ValidateCertificateChain/Async() private for now
These might be valuable to make public at some point, but when we do,
we might want to provide more info as to why they did not validate.

Unfortunately, BouncyCastle does not make that easy to programatically
determine via their APIs, so maybe we'd have to live with exceptions.
Either way, true/false might not be ideal for a public API.

Unit Test changes are all code formatting/style related changes.