Update MimeReader to use Span<T> for built-in intrinsics
It's gotten to a point where .NET Core's built-in SIMD optimizations
in Span<T>.IndexOf() have gotten as good or better than my current
optimization for scanning until a '\n' in ScanContent().

It also allows me to now use Span<T>.IndexOf() in all other locations
that were using `while (*inptr != '\n') inptr++`, so we get improved
performance in those locations for free.

In some ways, it's sad to see the end of an era - but also exciting
because MimeReader is now faster than ever!