Optimization: Use currentBoundary state to skip over boundary markers
Instead of having MimeReader.SkipLine/Async() loop over the entire line
containing what we know is a boundary marker, just skip over the known
length of the currentBoundary - this way we only need to loop over the
trailing whitespace characters.

Also, there's no need to call ReadAhead because we know we already have
the full line.