Modified the ExperimentalMimeParser to record raw boundary markers
MimeParser and previous iterations of the ExperimentalMimeParser did not
record the exact boundary marker used in the MIME stream to separate each
child of a multipart.

Normally, this isn't a problem because the boundary marker used by properly
written email software is just "--{boundary}\r\n", BUT according to the
MIME specifications, these lines are allowed to contain trailing whitespace
characters that should be ignored by MIME parsers (which is what MimeParser
and ExperimentalMimeParser have been doing).

That said, when writing a Multipart back out to a stream, if we do not
record the *exact* boundary markers used between each child part, then we
cannot write out an *exact* copy of what the parser encountered in the
original MIME stream.

As far as I have been able to tell, this is an extremely uncommon scenario
but in the interest of persuing perfection, ExperimentalMimeParser and
Multipart have now been modified to handle this.

MimeParser still suffers from this flaw, but I'll be looking into solving
that next.