Improve previous fix to avoid endless loops if a line is extremely long
(extremely long meaning the length exceeds the buffer length)

We do this by assuming that any line >= 1000 octets is definitely *not* a
boundary marker.

Also added other smarts to consume partial lines that can't possibly be a
boundary marker to avoid re-scanning over the same input bytes in the next
pass.