Modified AsymmetricAlgorithmExtensions to satisfy static analysis
When instantiating DSACryptoServiceProvider or RSACryptoServiceProvider,
use a minimum key size value of 2048 in order to satisfy static analysis
tools.

Technically, this is not needed because we import the key information that
we exported from BouncyCastle, so this isn't generating a *new* DSA or
RSA key which means the key size is effectively ignored but it satisfies
static analysis tools that don't understand this.