Use MemoryStreams in MimeParser benchmarks
This way we eliminate disk I/O inconsistencies between runs and should,
at least in theory, get more accurate comparison data.

In fairness, the downside of this is that any performance gains or losses
will be amplified compared to normal usage (which I assume will be parsing
FileStreams as opposed to MemoryStreams).