Refactored MimeReader to use ReadOnlySpan<byte> instead of unsafe byte*
The main advantage of this change is to make the code a bit nicer. The
AsyncMimeReader code, especially, becomes a lot nicer with this change.

The other advantage is that Spans don't interfere with the GC like pinned
memory buffers do, which could theoretically mean that using MimeReader is
less likely to result in memory fragmentation (although not sure on the
real-world consequences of the previous code in practice).

The downside is that ReadOnlySpan<byte>.IndexOf() is *slower* than the
current implementation on .NET Framework and possibly also even .NET Core
where the platform architecture does not have support for SIMD.

That's why this PR is a Work-In-Progress and not yet merged to master.