Optimized the Base64Encoder to use hardware acceleration when available (#1176)
* Optimized the Base64Encoder to use hardware acceleration when available

Also replaced Base64Encoder for rfc2047 encoded-word uses with a new
BEncoder class that doesn't need to keep state.

* Make the non-hwaccel Base64Encoder code-path unit testable

* Improved encoder/decoder benchmarks a bit

* Don't cap Base64Encoder's maxLineLength to 76

After some benchmarking, it seems that a maxLineLength of 80 is pretty close
to the maximum throughput that we can get out of the hwaccel code-paths
(at least on a CPU that only supports up to AVX2) without going
significantly over the MIME standard line length.

While maxLineLengths of 96, 112, and 128 are all faster (for example),
they are barely faster.

That said, the Base64.EncodeToUtf8() method is still, obviously, faster.