Added BinaryQuotedPrintableDecoder
Some mail software will attach pdf files using the quoted-printable
encoding.

On Linux/macOS, because MimeKit saves messages with the Unix line endings,
decoding these pdfs can produce corrupted output when the pdf contains
\r's of significance. These same pdfs typically decode correctly on Windows
because it just so happens that unencoded \r's aren't stripped on that
platform.

BinaryQuotedPrintableDecoder re-inserts the \r's as needed during the
decoding process to overcome this problem.

Note: The current unit test does not properly test this scenario and so
needs some work. Unfortunately, none of the pdfs that I have received that
illustrate this problem can be made public.